@@ -1,20 +1,21 @@
 [package]
 name = "serde"
-version = "1.0.175" # remember to update html_root_url and serde_derive dependency
+version = "1.0.181" # remember to update html_root_url and serde_derive dependency
 authors = ["Erick Tryzelaar <erick.tryzelaar@gmail.com>", "David Tolnay <dtolnay@gmail.com>"]
 build = "build.rs"
 categories = ["encoding", "no-std", "no-std::no-alloc"]
 description = "A generic serialization/deserialization framework"
 documentation = "https://docs.rs/serde"
+edition = "2018"
 homepage = "https://serde.rs"
 keywords = ["serde", "serialization", "no_std"]
 license = "MIT OR Apache-2.0"
 readme = "crates-io.md"
 repository = "https://github.com/serde-rs/serde"
-rust-version = "1.19"
+rust-version = "1.31"
 
 [dependencies]
-serde_derive = { version = "=1.0.175", optional = true, path = "../serde_derive" }
+serde_derive = { version = "=1.0.181", optional = true, path = "../serde_derive" }
 
 [dev-dependencies]
 serde_derive = { version = "1", path = "../serde_derive" }
