@@ -13,10 +13,7 @@
 //!
 //! [https://serde.rs/derive.html]: https://serde.rs/derive.html
 
-#![doc(html_root_url = "https://docs.rs/serde_derive/1.0.175")]
-#![allow(unknown_lints, bare_trait_objects)]
-
-extern crate proc_macro;
+#![doc(html_root_url = "https://docs.rs/serde_derive/1.0.181")]
 
 #[cfg(not(all(target_arch = "x86_64", target_os = "linux", target_env = "gnu")))]
 include!("lib_from_source.rs");
