@@ -1,16 +1,14 @@
-use lib::*;
+use crate::lib::*;
 
-use de::{
-    Deserialize, Deserializer, EnumAccess, Error, SeqAccess, Unexpected, VariantAccess, Visitor,
+use crate::de::{
+    Deserialize, Deserializer, EnumAccess, Error, MapAccess, SeqAccess, Unexpected, VariantAccess,
+    Visitor,
 };
 
-#[cfg(any(feature = "std", feature = "alloc", not(no_core_duration)))]
-use de::MapAccess;
-
-use seed::InPlaceSeed;
+use crate::seed::InPlaceSeed;
 
 #[cfg(any(feature = "std", feature = "alloc"))]
-use __private::size_hint;
+use crate::de::size_hint;
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -84,7 +82,7 @@ macro_rules! impl_deserialize_num {
     ($primitive:ident, $nonzero:ident $(cfg($($cfg:tt)*))*, $deserialize:ident $($method:ident!($($val:ident : $visit:ident)*);)*) => {
         impl_deserialize_num!($primitive, $deserialize $($method!($($val : $visit)*);)*);
 
-        #[cfg(all(not(no_num_nonzero), $($($cfg)*)*))]
+        $(#[cfg($($cfg)*)])*
         impl<'de> Deserialize<'de> for num::$nonzero {
             fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
             where
@@ -684,7 +682,7 @@ impl<'de> Visitor<'de> for CStringVisitor {
         let capacity = size_hint::cautious::<u8>(seq.size_hint());
         let mut values = Vec::<u8>::with_capacity(capacity);
 
-        while let Some(value) = try!(seq.next_element()) {
+        while let Some(value) = tri!(seq.next_element()) {
             values.push(value);
         }
 
@@ -747,13 +745,9 @@ macro_rules! forwarded_impl {
     }
 }
 
-#[cfg(all(
-    any(feature = "std", all(not(no_core_cstr), feature = "alloc")),
-    not(no_de_boxed_c_str)
-))]
+#[cfg(any(feature = "std", all(not(no_core_cstr), feature = "alloc")))]
 forwarded_impl!((), Box<CStr>, CString::into_boxed_c_str);
 
-#[cfg(not(no_core_reverse))]
 forwarded_impl!((T), Reverse<T>, Reverse);
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -901,7 +895,7 @@ macro_rules! seq_impl {
                     {
                         let mut values = $with_capacity;
 
-                        while let Some(value) = try!($access.next_element()) {
+                        while let Some(value) = tri!($access.next_element()) {
                             $insert(&mut values, value);
                         }
 
@@ -939,7 +933,7 @@ macro_rules! seq_impl {
                         $reserve(&mut self.0, size_hint::cautious::<T>($access.size_hint()));
 
                         // FIXME: try to overwrite old values here? (Vec, VecDeque, LinkedList)
-                        while let Some(value) = try!($access.next_element()) {
+                        while let Some(value) = tri!($access.next_element()) {
                             $insert(&mut self.0, value);
                         }
 
@@ -1039,7 +1033,7 @@ where
                 let capacity = size_hint::cautious::<T>(seq.size_hint());
                 let mut values = Vec::<T>::with_capacity(capacity);
 
-                while let Some(value) = try!(seq.next_element()) {
+                while let Some(value) = tri!(seq.next_element()) {
                     values.push(value);
                 }
 
@@ -1081,7 +1075,7 @@ where
                 for i in 0..self.0.len() {
                     let next = {
                         let next_place = InPlaceSeed(&mut self.0[i]);
-                        try!(seq.next_element_seed(next_place))
+                        tri!(seq.next_element_seed(next_place))
                     };
                     if next.is_none() {
                         self.0.truncate(i);
@@ -1089,7 +1083,7 @@ where
                     }
                 }
 
-                while let Some(value) = try!(seq.next_element()) {
+                while let Some(value) = tri!(seq.next_element()) {
                     self.0.push(value);
                 }
 
@@ -1161,7 +1155,7 @@ macro_rules! array_impls {
                     A: SeqAccess<'de>,
                 {
                     Ok([$(
-                        match try!(seq.next_element()) {
+                        match tri!(seq.next_element()) {
                             Some(val) => val,
                             None => return Err(Error::invalid_length($n, &self)),
                         }
@@ -1186,7 +1180,7 @@ macro_rules! array_impls {
                 {
                     let mut fail_idx = None;
                     for (idx, dest) in self.0[..].iter_mut().enumerate() {
-                        if try!(seq.next_element_seed(InPlaceSeed(dest))).is_none() {
+                        if tri!(seq.next_element_seed(InPlaceSeed(dest))).is_none() {
                             fail_idx = Some(idx);
                             break;
                         }
@@ -1284,7 +1278,7 @@ macro_rules! tuple_impls {
                             A: SeqAccess<'de>,
                         {
                             $(
-                                let $name = match try!(seq.next_element()) {
+                                let $name = match tri!(seq.next_element()) {
                                     Some(value) => value,
                                     None => return Err(Error::invalid_length($n, &self)),
                                 };
@@ -1318,7 +1312,7 @@ macro_rules! tuple_impls {
                             A: SeqAccess<'de>,
                         {
                             $(
-                                if try!(seq.next_element_seed(InPlaceSeed(&mut (self.0).$n))).is_none() {
+                                if tri!(seq.next_element_seed(InPlaceSeed(&mut (self.0).$n))).is_none() {
                                     return Err(Error::invalid_length($n, &self));
                                 }
                             )+
@@ -1395,7 +1389,7 @@ macro_rules! map_impl {
                     {
                         let mut values = $with_capacity;
 
-                        while let Some((key, value)) = try!($access.next_entry()) {
+                        while let Some((key, value)) = tri!($access.next_entry()) {
                             values.insert(key, value);
                         }
 
@@ -1541,7 +1535,7 @@ macro_rules! deserialize_enum {
             where
                 A: EnumAccess<'de>,
             {
-                match try!(data.variant()) {
+                match tri!(data.variant()) {
                     $(
                         ($name_kind :: $variant, v) => v.newtype_variant().map($name :: $variant),
                     )*
@@ -1561,7 +1555,7 @@ impl<'de> Deserialize<'de> for net::IpAddr {
         if deserializer.is_human_readable() {
             deserializer.deserialize_str(FromStrVisitor::new("IP address"))
         } else {
-            use lib::net::IpAddr;
+            use crate::lib::net::IpAddr;
             deserialize_enum! {
                 IpAddr IpAddrKind (V4; b"V4"; 0, V6; b"V6"; 1)
                 "`V4` or `V6`",
@@ -1604,7 +1598,7 @@ impl<'de> Deserialize<'de> for net::SocketAddr {
         if deserializer.is_human_readable() {
             deserializer.deserialize_str(FromStrVisitor::new("socket address"))
         } else {
-            use lib::net::SocketAddr;
+            use crate::lib::net::SocketAddr;
             deserialize_enum! {
                 SocketAddr SocketAddrKind (V4; b"V4"; 0, V6; b"V6"; 1)
                 "`V4` or `V6`",
@@ -1714,7 +1708,7 @@ impl<'de> Deserialize<'de> for PathBuf {
     }
 }
 
-#[cfg(all(feature = "std", not(no_de_boxed_path)))]
+#[cfg(feature = "std")]
 forwarded_impl!((), Box<Path>, PathBuf::into_boxed_path);
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -1748,7 +1742,7 @@ impl<'de> Visitor<'de> for OsStringVisitor {
     {
         use std::os::unix::ffi::OsStringExt;
 
-        match try!(data.variant()) {
+        match tri!(data.variant()) {
             (OsStringKind::Unix, v) => v.newtype_variant().map(OsString::from_vec),
             (OsStringKind::Windows, _) => Err(Error::custom(
                 "cannot deserialize Windows OS string on Unix",
@@ -1763,7 +1757,7 @@ impl<'de> Visitor<'de> for OsStringVisitor {
     {
         use std::os::windows::ffi::OsStringExt;
 
-        match try!(data.variant()) {
+        match tri!(data.variant()) {
             (OsStringKind::Windows, v) => v
                 .newtype_variant::<Vec<u16>>()
                 .map(|vec| OsString::from_wide(&vec)),
@@ -1795,30 +1789,6 @@ forwarded_impl!((T), Box<[T]>, Vec::into_boxed_slice);
 #[cfg(any(feature = "std", feature = "alloc"))]
 forwarded_impl!((), Box<str>, String::into_boxed_str);
 
-#[cfg(all(no_de_rc_dst, feature = "rc", any(feature = "std", feature = "alloc")))]
-forwarded_impl! {
-    /// This impl requires the [`"rc"`] Cargo feature of Serde.
-    ///
-    /// Deserializing a data structure containing `Arc` will not attempt to
-    /// deduplicate `Arc` references to the same data. Every deserialized `Arc`
-    /// will end up with a strong count of 1.
-    ///
-    /// [`"rc"`]: https://serde.rs/feature-flags.html#-features-rc
-    (T), Arc<T>, Arc::new
-}
-
-#[cfg(all(no_de_rc_dst, feature = "rc", any(feature = "std", feature = "alloc")))]
-forwarded_impl! {
-    /// This impl requires the [`"rc"`] Cargo feature of Serde.
-    ///
-    /// Deserializing a data structure containing `Rc` will not attempt to
-    /// deduplicate `Rc` references to the same data. Every deserialized `Rc`
-    /// will end up with a strong count of 1.
-    ///
-    /// [`"rc"`]: https://serde.rs/feature-flags.html#-features-rc
-    (T), Rc<T>, Rc::new
-}
-
 #[cfg(any(feature = "std", feature = "alloc"))]
 impl<'de, 'a, T: ?Sized> Deserialize<'de> for Cow<'a, T>
 where
@@ -1849,7 +1819,7 @@ where
     where
         D: Deserializer<'de>,
     {
-        try!(Option::<T>::deserialize(deserializer));
+        tri!(Option::<T>::deserialize(deserializer));
         Ok(RcWeak::new())
     }
 }
@@ -1867,18 +1837,14 @@ where
     where
         D: Deserializer<'de>,
     {
-        try!(Option::<T>::deserialize(deserializer));
+        tri!(Option::<T>::deserialize(deserializer));
         Ok(ArcWeak::new())
     }
 }
 
 ////////////////////////////////////////////////////////////////////////////////
 
-#[cfg(all(
-    not(no_de_rc_dst),
-    feature = "rc",
-    any(feature = "std", feature = "alloc")
-))]
+#[cfg(all(feature = "rc", any(feature = "std", feature = "alloc")))]
 macro_rules! box_forwarded_impl {
     (
         $(#[doc = $doc:tt])*
@@ -1899,11 +1865,7 @@ macro_rules! box_forwarded_impl {
     };
 }
 
-#[cfg(all(
-    not(no_de_rc_dst),
-    feature = "rc",
-    any(feature = "std", feature = "alloc")
-))]
+#[cfg(all(feature = "rc", any(feature = "std", feature = "alloc")))]
 box_forwarded_impl! {
     /// This impl requires the [`"rc"`] Cargo feature of Serde.
     ///
@@ -1915,11 +1877,7 @@ box_forwarded_impl! {
     Rc
 }
 
-#[cfg(all(
-    not(no_de_rc_dst),
-    feature = "rc",
-    any(feature = "std", feature = "alloc")
-))]
+#[cfg(all(feature = "rc", any(feature = "std", feature = "alloc")))]
 box_forwarded_impl! {
     /// This impl requires the [`"rc"`] Cargo feature of Serde.
     ///
@@ -1963,7 +1921,6 @@ forwarded_impl!((T), RwLock<T>, RwLock::new);
 //         secs: u64,
 //         nanos: u32,
 //     }
-#[cfg(any(feature = "std", not(no_core_duration)))]
 impl<'de> Deserialize<'de> for Duration {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
     where
@@ -2011,7 +1968,7 @@ impl<'de> Deserialize<'de> for Duration {
                             b"secs" => Ok(Field::Secs),
                             b"nanos" => Ok(Field::Nanos),
                             _ => {
-                                let value = ::__private::from_utf8_lossy(value);
+                                let value = crate::__private::from_utf8_lossy(value);
                                 Err(Error::unknown_field(&*value, FIELDS))
                             }
                         }
@@ -2046,19 +2003,19 @@ impl<'de> Deserialize<'de> for Duration {
             where
                 A: SeqAccess<'de>,
             {
-                let secs: u64 = match try!(seq.next_element()) {
+                let secs: u64 = match tri!(seq.next_element()) {
                     Some(value) => value,
                     None => {
                         return Err(Error::invalid_length(0, &self));
                     }
                 };
-                let nanos: u32 = match try!(seq.next_element()) {
+                let nanos: u32 = match tri!(seq.next_element()) {
                     Some(value) => value,
                     None => {
                         return Err(Error::invalid_length(1, &self));
                     }
                 };
-                try!(check_overflow(secs, nanos));
+                tri!(check_overflow(secs, nanos));
                 Ok(Duration::new(secs, nanos))
             }
 
@@ -2068,19 +2025,19 @@ impl<'de> Deserialize<'de> for Duration {
             {
                 let mut secs: Option<u64> = None;
                 let mut nanos: Option<u32> = None;
-                while let Some(key) = try!(map.next_key()) {
+                while let Some(key) = tri!(map.next_key()) {
                     match key {
                         Field::Secs => {
                             if secs.is_some() {
                                 return Err(<A::Error as Error>::duplicate_field("secs"));
                             }
-                            secs = Some(try!(map.next_value()));
+                            secs = Some(tri!(map.next_value()));
                         }
                         Field::Nanos => {
                             if nanos.is_some() {
                                 return Err(<A::Error as Error>::duplicate_field("nanos"));
                             }
-                            nanos = Some(try!(map.next_value()));
+                            nanos = Some(tri!(map.next_value()));
                         }
                     }
                 }
@@ -2092,7 +2049,7 @@ impl<'de> Deserialize<'de> for Duration {
                     Some(nanos) => nanos,
                     None => return Err(<A::Error as Error>::missing_field("nanos")),
                 };
-                try!(check_overflow(secs, nanos));
+                tri!(check_overflow(secs, nanos));
                 Ok(Duration::new(secs, nanos))
             }
         }
@@ -2184,19 +2141,19 @@ impl<'de> Deserialize<'de> for SystemTime {
             where
                 A: SeqAccess<'de>,
             {
-                let secs: u64 = match try!(seq.next_element()) {
+                let secs: u64 = match tri!(seq.next_element()) {
                     Some(value) => value,
                     None => {
                         return Err(Error::invalid_length(0, &self));
                     }
                 };
-                let nanos: u32 = match try!(seq.next_element()) {
+                let nanos: u32 = match tri!(seq.next_element()) {
                     Some(value) => value,
                     None => {
                         return Err(Error::invalid_length(1, &self));
                     }
                 };
-                try!(check_overflow(secs, nanos));
+                tri!(check_overflow(secs, nanos));
                 Ok(Duration::new(secs, nanos))
             }
 
@@ -2206,7 +2163,7 @@ impl<'de> Deserialize<'de> for SystemTime {
             {
                 let mut secs: Option<u64> = None;
                 let mut nanos: Option<u32> = None;
-                while let Some(key) = try!(map.next_key()) {
+                while let Some(key) = tri!(map.next_key()) {
                     match key {
                         Field::Secs => {
                             if secs.is_some() {
@@ -2214,7 +2171,7 @@ impl<'de> Deserialize<'de> for SystemTime {
                                     "secs_since_epoch",
                                 ));
                             }
-                            secs = Some(try!(map.next_value()));
+                            secs = Some(tri!(map.next_value()));
                         }
                         Field::Nanos => {
                             if nanos.is_some() {
@@ -2222,7 +2179,7 @@ impl<'de> Deserialize<'de> for SystemTime {
                                     "nanos_since_epoch",
                                 ));
                             }
-                            nanos = Some(try!(map.next_value()));
+                            nanos = Some(tri!(map.next_value()));
                         }
                     }
                 }
@@ -2234,13 +2191,13 @@ impl<'de> Deserialize<'de> for SystemTime {
                     Some(nanos) => nanos,
                     None => return Err(<A::Error as Error>::missing_field("nanos_since_epoch")),
                 };
-                try!(check_overflow(secs, nanos));
+                tri!(check_overflow(secs, nanos));
                 Ok(Duration::new(secs, nanos))
             }
         }
 
         const FIELDS: &[&str] = &["secs_since_epoch", "nanos_since_epoch"];
-        let duration = try!(deserializer.deserialize_struct("SystemTime", FIELDS, DurationVisitor));
+        let duration = tri!(deserializer.deserialize_struct("SystemTime", FIELDS, DurationVisitor));
         #[cfg(not(no_systemtime_checked_add))]
         let ret = UNIX_EPOCH
             .checked_add(duration)
@@ -2269,7 +2226,7 @@ where
     where
         D: Deserializer<'de>,
     {
-        let (start, end) = try!(deserializer.deserialize_struct(
+        let (start, end) = tri!(deserializer.deserialize_struct(
             "Range",
             range::FIELDS,
             range::RangeVisitor {
@@ -2281,7 +2238,6 @@ where
     }
 }
 
-#[cfg(not(no_range_inclusive))]
 impl<'de, Idx> Deserialize<'de> for RangeInclusive<Idx>
 where
     Idx: Deserialize<'de>,
@@ -2290,7 +2246,7 @@ where
     where
         D: Deserializer<'de>,
     {
-        let (start, end) = try!(deserializer.deserialize_struct(
+        let (start, end) = tri!(deserializer.deserialize_struct(
             "RangeInclusive",
             range::FIELDS,
             range::RangeVisitor {
@@ -2303,9 +2259,9 @@ where
 }
 
 mod range {
-    use lib::*;
+    use crate::lib::*;
 
-    use de::{Deserialize, Deserializer, Error, MapAccess, SeqAccess, Visitor};
+    use crate::de::{Deserialize, Deserializer, Error, MapAccess, SeqAccess, Visitor};
 
     pub const FIELDS: &[&str] = &["start", "end"];
 
@@ -2351,7 +2307,7 @@ mod range {
                         b"start" => Ok(Field::Start),
                         b"end" => Ok(Field::End),
                         _ => {
-                            let value = ::__private::from_utf8_lossy(value);
+                            let value = crate::__private::from_utf8_lossy(value);
                             Err(Error::unknown_field(&*value, FIELDS))
                         }
                     }
@@ -2381,13 +2337,13 @@ mod range {
         where
             A: SeqAccess<'de>,
         {
-            let start: Idx = match try!(seq.next_element()) {
+            let start: Idx = match tri!(seq.next_element()) {
                 Some(value) => value,
                 None => {
                     return Err(Error::invalid_length(0, &self));
                 }
             };
-            let end: Idx = match try!(seq.next_element()) {
+            let end: Idx = match tri!(seq.next_element()) {
                 Some(value) => value,
                 None => {
                     return Err(Error::invalid_length(1, &self));
@@ -2402,19 +2358,19 @@ mod range {
         {
             let mut start: Option<Idx> = None;
             let mut end: Option<Idx> = None;
-            while let Some(key) = try!(map.next_key()) {
+            while let Some(key) = tri!(map.next_key()) {
                 match key {
                     Field::Start => {
                         if start.is_some() {
                             return Err(<A::Error as Error>::duplicate_field("start"));
                         }
-                        start = Some(try!(map.next_value()));
+                        start = Some(tri!(map.next_value()));
                     }
                     Field::End => {
                         if end.is_some() {
                             return Err(<A::Error as Error>::duplicate_field("end"));
                         }
-                        end = Some(try!(map.next_value()));
+                        end = Some(tri!(map.next_value()));
                     }
                 }
             }
@@ -2448,7 +2404,7 @@ where
     where
         D: Deserializer<'de>,
     {
-        let start = try!(deserializer.deserialize_struct(
+        let start = tri!(deserializer.deserialize_struct(
             "RangeFrom",
             range_from::FIELDS,
             range_from::RangeFromVisitor {
@@ -2461,9 +2417,9 @@ where
 }
 
 mod range_from {
-    use lib::*;
+    use crate::lib::*;
 
-    use de::{Deserialize, Deserializer, Error, MapAccess, SeqAccess, Visitor};
+    use crate::de::{Deserialize, Deserializer, Error, MapAccess, SeqAccess, Visitor};
 
     pub const FIELDS: &[&str] = &["end"];
 
@@ -2506,7 +2462,7 @@ mod range_from {
                     match value {
                         b"end" => Ok(Field::End),
                         _ => {
-                            let value = ::__private::from_utf8_lossy(value);
+                            let value = crate::__private::from_utf8_lossy(value);
                             Err(Error::unknown_field(&*value, FIELDS))
                         }
                     }
@@ -2536,7 +2492,7 @@ mod range_from {
         where
             A: SeqAccess<'de>,
         {
-            let end: Idx = match try!(seq.next_element()) {
+            let end: Idx = match tri!(seq.next_element()) {
                 Some(value) => value,
                 None => {
                     return Err(Error::invalid_length(0, &self));
@@ -2550,13 +2506,13 @@ mod range_from {
             A: MapAccess<'de>,
         {
             let mut end: Option<Idx> = None;
-            while let Some(key) = try!(map.next_key()) {
+            while let Some(key) = tri!(map.next_key()) {
                 match key {
                     Field::End => {
                         if end.is_some() {
                             return Err(<A::Error as Error>::duplicate_field("end"));
                         }
-                        end = Some(try!(map.next_value()));
+                        end = Some(tri!(map.next_value()));
                     }
                 }
             }
@@ -2586,7 +2542,7 @@ where
     where
         D: Deserializer<'de>,
     {
-        let end = try!(deserializer.deserialize_struct(
+        let end = tri!(deserializer.deserialize_struct(
             "RangeTo",
             range_to::FIELDS,
             range_to::RangeToVisitor {
@@ -2599,9 +2555,9 @@ where
 }
 
 mod range_to {
-    use lib::*;
+    use crate::lib::*;
 
-    use de::{Deserialize, Deserializer, Error, MapAccess, SeqAccess, Visitor};
+    use crate::de::{Deserialize, Deserializer, Error, MapAccess, SeqAccess, Visitor};
 
     pub const FIELDS: &[&str] = &["start"];
 
@@ -2644,7 +2600,7 @@ mod range_to {
                     match value {
                         b"start" => Ok(Field::Start),
                         _ => {
-                            let value = ::__private::from_utf8_lossy(value);
+                            let value = crate::__private::from_utf8_lossy(value);
                             Err(Error::unknown_field(&*value, FIELDS))
                         }
                     }
@@ -2674,7 +2630,7 @@ mod range_to {
         where
             A: SeqAccess<'de>,
         {
-            let start: Idx = match try!(seq.next_element()) {
+            let start: Idx = match tri!(seq.next_element()) {
                 Some(value) => value,
                 None => {
                     return Err(Error::invalid_length(0, &self));
@@ -2688,13 +2644,13 @@ mod range_to {
             A: MapAccess<'de>,
         {
             let mut start: Option<Idx> = None;
-            while let Some(key) = try!(map.next_key()) {
+            while let Some(key) = tri!(map.next_key()) {
                 match key {
                     Field::Start => {
                         if start.is_some() {
                             return Err(<A::Error as Error>::duplicate_field("start"));
                         }
-                        start = Some(try!(map.next_value()));
+                        start = Some(tri!(map.next_value()));
                     }
                 }
             }
@@ -2709,7 +2665,6 @@ mod range_to {
 
 ////////////////////////////////////////////////////////////////////////////////
 
-#[cfg(any(not(no_ops_bound), all(feature = "std", not(no_collections_bound))))]
 impl<'de, T> Deserialize<'de> for Bound<T>
 where
     T: Deserialize<'de>,
@@ -2801,7 +2756,7 @@ where
             where
                 A: EnumAccess<'de>,
             {
-                match try!(data.variant()) {
+                match tri!(data.variant()) {
                     (Field::Unbounded, v) => v.unit_variant().map(|()| Bound::Unbounded),
                     (Field::Included, v) => v.newtype_variant().map(Bound::Included),
                     (Field::Excluded, v) => v.newtype_variant().map(Bound::Excluded),
@@ -2910,7 +2865,7 @@ where
             where
                 A: EnumAccess<'de>,
             {
-                match try!(data.variant()) {
+                match tri!(data.variant()) {
                     (Field::Ok, v) => v.newtype_variant().map(Ok),
                     (Field::Err, v) => v.newtype_variant().map(Err),
                 }
