@@ -1,11 +1,7 @@
 use proc_macro2::TokenStream;
-
-use syn;
-use try;
+use quote::quote;
 
 pub fn wrap_in_const(serde_path: Option<&syn::Path>, code: TokenStream) -> TokenStream {
-    let try_replacement = try::replacement();
-
     let use_serde = match serde_path {
         Some(path) => quote! {
             use #path as _serde;
@@ -21,7 +17,6 @@ pub fn wrap_in_const(serde_path: Option<&syn::Path>, code: TokenStream) -> Token
         #[allow(non_upper_case_globals, unused_attributes, unused_qualifications)]
         const _: () = {
             #use_serde
-            #try_replacement
             #code
         };
     }
